# openapi.yaml
openapi: 3.0.3
info:
  title: Nova Stream API
  version: "0.1.0"
servers:
  - url: https://localhost:3000/api
paths:
  /auth/register:
    post:
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        "200": { description: OK }
  /auth/login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        "200": { description: OK }
  /auth/refresh:
    post:
      summary: Refresh access token
      responses:
        "200": { description: OK }
  /xtream/link:
    post:
      summary: Link Xtream account (encrypted at rest)
      responses:
        "200": { description: OK }
  /xtream/detail:
    get:
      summary: Get content detail (proxied)
      parameters:
        - name: type
          in: query
          schema: { type: string, enum: [movie, series, episode, live] }
        - name: id
          in: query
          schema: { type: string }
      responses:
        "200": { description: OK }
  /xtream/stream:
    get:
      summary: Get stream URL for player
      parameters:
        - name: type
          in: query
          schema: { type: string, enum: [movie, series, episode, live] }
        - name: id
          in: query
          schema: { type: string }
      responses:
        "200": { description: OK }
  /tmdb/review:
    get:
      summary: Get TMDB overview by title
      parameters:
        - name: type
          in: query
          schema: { type: string, enum: [movie, series] }
        - name: title
          in: query
          schema: { type: string }
      responses:
        "200": { description: OK }